Creates separate routes file for spotify routes.
