+++ index.js
@@ -6,7 +6,7 @@ const app = express();
 app.listen(3000, () => console.log('Example app listening on port 3000!'));
 app.use(express.static(path.join(__dirname, '/public')));
 
-app.use('/spotify', require('./routes/routes'));
+app.use('/spotify', require('./routes/spotify'));
 
 require('./utils/sendgrid');
